+++ sleek_storage/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.1.0
 * BREAKING: `fromJson` now also provides the key.
+* New `Box.containsKey` method.
 
 ## 1.0.0
 * First stable release.